storing padded string is the same as storing string-padded blob
and means we can skip the casting to base36, which is a little confusing.
originally had hoped to store the sort_order value as an integer, but sqlite
doesn't have a numeric type large enough to store these, and once we're storing
as a BLOB lost the benefit of storing as integer anyway. so might as well
drop a lot of unneeded code and store a more human-readable sort value.